Update SVG color scheme in avatar tests
The color values in the SVG output of the Larvatar and InitialsAvatar classes have been modified in the test scripts. These changes reflect the updated color scheme for avatar generation - from a previous pink (#e8d2dc) to a lighter version (#f3dde7), and from a dark red (#690230) to a brighter purple (#9c3563).