Add support for different avatar forms
The main changes revolve around adding the option to generate either a circle or a square form for the avatar. This is accomplished by introducing a new FormTypes Enum and allowing the form type to be set and fetched. The generate function has been altered to adapt to these changes and to generate an SVG of the correct form based on the form type selected.