Update font properties and refactor initials retrieval
The fontWeight property of InitialsAvatar class was changed from 'normal' to 'regular'. Additionally, a new method 'getName' was added to get the name of the avatar. Lastly, the method 'getInitials' was refactored to directly get initials from the name property instead of an array of names, and the fontFamily hard-coded to "Segoe UI, Helvetica, sans-serif".