Refactor Larvatar class and modify constructor
The Larvatar class was refactored for better organization and readability. The constructor's signature was changed to take a mandatory LarvatarTypes object instead of a mix-type parameter. A static `make` method was added for more intuitive object creation. The `$name` property is now an instance of the `Name` class, and the method `getBase64()` was moved towards the end of the class for logical grouping.